+++ src/net/stun/message.cpp
@@ -1,27 +1,23 @@
-#include <cassert>
-
-#include "byteorder.hpp"
 #include "message.hpp"
 
+#include "byteorder.hpp"
+
+#include <cassert> // assert
+#include <cstring> // std::memcpy
 
 namespace shar::net::stun {
 
-bool is_message(const u8* data, usize size) noexcept {
+bool is_message(const u8 *data, usize size) noexcept {
   // safety: no code below uses non-const methods
-  Message message{const_cast<u8*>(data), size};
-  return message.valid() &&
-        (message.message_type() >> 14) == 0 &&
-        (message.length() & 0b11) == 0 &&
-        message.cookie() == Message::MAGIC;
+  Message message{const_cast<u8 *>(data), size};
+  return message.valid() && (message.message_type() >> 14) == 0 &&
+         (message.length() & 0b11) == 0 && message.cookie() == Message::MAGIC;
 }
 
-Message::Message(u8* data, usize size) noexcept
-  : BytesRef(data, size)
-{}
+Message::Message(u8 *data, usize size) noexcept : BytesRef(data, size) {}
 
 bool Message::valid() const noexcept {
-  return m_data != nullptr &&
-         m_size >= MIN_SIZE;
+  return m_data != nullptr && m_size >= MIN_SIZE;
 }
 
 u16 Message::message_type() const noexcept {
@@ -39,9 +35,9 @@ u16 Message::method() const noexcept {
   assert(valid());
   u8 msb = m_data[0];
   u8 lsb = m_data[1];
-  return ((u16{msb} & 0b00111110) << 7) |
-         ((u16{lsb} & 0b11100000) >> 1) |
-         (u16{lsb} & 0b00001111);
+  return static_cast<u16>(((u16{msb} & 0b00111110) << 7) |
+                          ((u16{lsb} & 0b11100000) >> 1) |
+                          (u16{lsb} & 0b00001111));
 }
 
 void Message::set_method(u16 m) noexcept {
@@ -64,7 +60,7 @@ u8 Message::type() const noexcept {
   assert(valid());
   u8 msb = m_data[0];
   u8 lsb = m_data[1];
-  return ((msb & 1) << 1) | ((lsb & (1 << 4)) >> 4);
+  return static_cast<u8>(((msb & 1) << 1) | ((lsb & (1 << 4)) >> 4));
 }
 
 void Message::set_type(u8 t) noexcept {
@@ -117,11 +113,11 @@ void Message::set_transaction(Transaction t) noexcept {
   std::memcpy(m_data + 8, t.data(), t.size());
 }
 
-u8* Message::payload() noexcept {
+u8 *Message::payload() noexcept {
   return m_data + MIN_SIZE;
 }
 
-const u8* Message::payload() const noexcept {
+const u8 *Message::payload() const noexcept {
   return m_data + MIN_SIZE;
 }
 
@@ -129,4 +125,4 @@ usize Message::payload_size() const noexcept {
   return length();
 }
 
-}
+} // namespace shar::net::stun
